requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
coverage==5.3.1
distlib==0.3.1
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
drf-nested-routers==0.92.5
filelock==3.0.12
future==0.18.2
importlib-metadata==1.7.0
iniconfig==1.1.1
model-bakery==1.2.1
packaging==20.8
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.1
pytest-cov==2.10.1
pytest-django==4.1.0
pytz==2020.4
serializers==0.2.4
six==1.12.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.0.31
zipp==3.1.0
